new commands, full database for this period
New commands:
"Lectures today/yesterday/tommorow"- obv
"Lectures in 5 days"  - gets the lectures in 5 days
"Lectures on this Friday" - Gets the lectures on the next friday from today, different from "Lectures on Friday" which returns every lecture that's on a Friday.
Also added every lecture for this period into the dataset.